Fix circular inclusion
