+++ yogit/storage/storage.py
@@ -5,6 +5,8 @@ yogit storage
 import os
 import yaml
 
+from yogit.yogit.logger import LOGGER
+
 
 class Storage:
     """ Storage based on YAML file """
@@ -25,8 +27,6 @@ class Storage:
             with open(self.filename, "r") as yaml_file:
                 return yaml.load(yaml_file, Loader=yaml.FullLoader) or {}
         except Exception as error:
-            from yogit.yogit.logger import LOGGER  # TODO fix mutual inclusion
-
             LOGGER.error(str(error))
             return {}
